requirements.txt
Flask==0.9
Flask-SQLAlchemy==0.16
-e git://github.com/kvesteri/flask-storage.git#egg=Flask-Storage
Jinja2==2.6
